moved more into pipe base class for any shared code
