+++ src/main/java/gtclassic/api/pipe/GTTilePipeFluidBase.java
@@ -1,13 +1,16 @@
 package gtclassic.api.pipe;
 
 import gtclassic.api.helpers.GTHelperFluid;
-import gtclassic.api.interfaces.IGTDebuggableTile;
+import gtclassic.common.GTLang;
 import ic2.core.fluid.IC2Tank;
+import ic2.core.inventory.container.ContainerIC2;
+import ic2.core.item.misc.ItemDisplayIcon;
+import ic2.core.platform.lang.components.base.LocaleComp;
 import ic2.core.util.obj.ITankListener;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.ITickable;
 import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.fluids.FluidStack;
@@ -16,8 +19,7 @@ import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 
-public abstract class GTTilePipeFluidBase extends GTTilePipeBase
-		implements ITankListener, IGTDebuggableTile, ITickable {
+public abstract class GTTilePipeFluidBase extends GTTilePipeBase implements ITankListener {
 
 	private IC2Tank tank;
 
@@ -25,6 +27,11 @@ public abstract class GTTilePipeFluidBase extends GTTilePipeBase
 		this.tank = new IC2Tank(capacity);
 		this.tank.addListener(this);
 		this.addGuiFields("tank");
+	}
+
+	@Override
+	public LocaleComp getBlockName() {
+		return GTLang.FLUID_PIPE_LANG;
 	}
 
 	@Override
@@ -40,6 +47,7 @@ public abstract class GTTilePipeFluidBase extends GTTilePipeBase
 
 	public void onTankChanged(IFluidTank tank) {
 		this.getNetwork().updateTileGuiField(this, "tank");
+		this.inventory.set(0, ItemDisplayIcon.createWithFluidStack(this.tank.getFluid()));
 	}
 
 	@Override
@@ -53,6 +61,11 @@ public abstract class GTTilePipeFluidBase extends GTTilePipeBase
 		super.writeToNBT(nbt);
 		this.tank.writeToNBT(this.getTag(nbt, "tank"));
 		return nbt;
+	}
+
+	@Override
+	public ContainerIC2 getGuiContainer(EntityPlayer player) {
+		return new GTContainerPipeFluid(player.inventory, this);
 	}
 
 	@Override
@@ -70,11 +83,6 @@ public abstract class GTTilePipeFluidBase extends GTTilePipeBase
 
 	public IC2Tank getTankInstance() {
 		return this.tank;
-	}
-
-	@Override
-	public boolean canUpdate() {
-		return this.isSimulating();
 	}
 
 	@Override
@@ -99,11 +107,6 @@ public abstract class GTTilePipeFluidBase extends GTTilePipeBase
 				}
 			}
 		}
-	}
-
-	@Override
-	public boolean canDebugWithMagnifyingGlass() {
-		return true;
 	}
 
 	@Override
