+++ src/main/resources/assets/gtclassic/lang/en_us.lang
@@ -71,7 +71,8 @@ tile.gtclassic.pipefluidbronze.name=Bronze Fluid Pipe
 tile.gtclassic.pipefluidinvar.name=Invar Fluid Pipe
 tile.gtclassic.pipefluidtitanium.name=Titanium Fluid Pipe
 tile.gtclassic.pipefluidhighpressure.name=High Pressure Fluid Pipe
-tile.gtclassic.pipename.name=Item Pipe
+tile.gtclassic.itempipename.name=Item Pipe
+tile.gtclassic.fluidpipename.name=Fluid Pipe
 
 #dust
 item.gtclassic.dustElectrum.name=Electrum Dust
